+++ adh_sample_library_preview/ADHClient.py
@@ -20,8 +20,8 @@ class ADHClient:
     A client that handles communication with AVEVA Data Hub
     """
 
-    def __init__(self, api_version: str, tenant: str, url: str, client_id: str = None,
-                 client_secret: str = None, accept_verbosity: bool = False, **kwargs):
+    def __init__(self, api_version: str, tenant: str, url: str, client_id: str,
+                 client_secret: str = None, accept_verbosity: bool = False, logging_enabled: bool = False, **kwargs):
         """
         Use this to help in communication with ADH
         :param api_version: Version of the api you are communicating with
@@ -31,13 +31,14 @@ class ADHClient:
         :param client_secret: Your client Secret or Key
         :param accept_verbosity: Sets whether in value calls you get all values or just
             non-default values
+        :param logging_enabled: Sets whether Python logging is enabled
         """
 
         if 'base_client' in kwargs:
             self.__base_client = kwargs.get('base_client')
         else:
             self.__base_client = BaseClient(api_version, tenant, url, client_id,
-                                            client_secret, accept_verbosity)
+                                            client_secret, accept_verbosity, logging_enabled)
 
         self.__asset_rules = AssetRules(self.__base_client)
         self.__assets = Assets(self.__base_client)
@@ -89,6 +90,17 @@ class ADHClient:
     @request_timeout.setter
     def request_timeout(self, value: int):
         self.__base_client.RequestTimeout = value
+
+    @property
+    def logging_enabled(self) -> bool:
+        """
+        :return: Whether logging is enabled (default False)
+        """
+        return self.__base_client.LoggingEnabled
+
+    @logging_enabled.setter
+    def logging_enabled(self, value: bool):
+        self.__base_client.LoggingEnabled = value
 
     @property
     def AssetRules(self) -> AssetRules: