+++ adh_sample_library_preview/EDSClient.py
@@ -10,12 +10,13 @@ class EDSClient:
     """
 
     def __init__(self, api_version: str = 'v1', url: str = 'http://localhost:5590',
-                 accept_verbosity: bool = False):
+                 accept_verbosity: bool = False, logging_enabled: bool = False):
         """
         Use this to help in communinication with EDS
         :param api_version: Version of the api you are communicating with, default is v1
         :param url: The base URL for EDS, default is http://localhost:5590
         :param accept_verbosity: Sets whether in value calls you get all values or just
+        :param logging_enabled: Sets whether Python logging is enabled
             non-default values
         """
         self.__base_client = BaseClient(
@@ -52,6 +53,17 @@ class EDSClient:
     @request_timeout.setter
     def request_timeout(self, value: int):
         self.__base_client.RequestTimeout = value
+
+    @property
+    def logging_enabled(self) -> bool:
+        """
+        :return: Whether logging is enabled (default False)
+        """
+        return self.__base_client.LoggingEnabled
+
+    @logging_enabled.setter
+    def logging_enabled(self, value: bool):
+        self.__base_client.LoggingEnabled = value
 
     @property
     def Types(self) -> Types: